==> jupyter_notebooks/01-DataLoading.ipynb <==
# **Notebook 1: Load data and data cleaning**

## Objectives

* Fetch data from Kaggle and save as raw data 
* Inspect data
* Save raw datasets
* Save cleaned data under outputs/datasets/cleaned

## Inputs

* Kaggle dataset downloaded from https://www.kaggle.com/datasets/codeinstitute/housing-prices-data

## Outputs

* Raw datasets are found in folder inputs/dataset/raw/house_prices
* Generate Dataset: outputs/datasets/collection/house_prices_records.csv
* Generate cleaned datasets: outputs/datasets/cleaned/house_prices_records_cleaned.csv and outputs/datasets/cleaned/inherited_houses_cleaned.csv

## Conclusions

* Data cleaning pipeline
* Drop variables: `['EnclosedPorch', 'WoodDeckSF' ]`
* Use median and mode imputation to replace missing values in numerical and categorical variables respectively.
* Handle mismatching data types by converting floats into integers



---

# Change working directory

* We are assuming you will store the notebooks in a subfolder, therefore when running the notebook in the editor, you will need to change the working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

# Section 1: Load and inspect the data

The data has been downloaded from: https://www.kaggle.com/datasets/codeinstitute/housing-prices-data

First we create the folders in the file path

import os
try:
  os.makedirs(name='inputs/datasets/raw/house_prices')
except Exception as e:
  print(e)

Then drag and drop both csv files to the folder. 
- house-metadata.txt can be dragged and dropped to inputs/datasets/raw for reference.

Loading the data and showing first rows to get an idea of the data I'm working with.

import pandas as pd

# Load the historical house prices records data
records_df = pd.read_csv('inputs/datasets/raw/house_prices/house_prices_records.csv')

# Load the inherited houses data
inherited_df = pd.read_csv('inputs/datasets/raw/house_prices/inherited_houses.csv')

# Check the first few rows of each DataFrame
print(records_df.head())

print(inherited_df.head())

# Section 2: Data overview

Explore the data to spot any anomalies. First let's take a look at records_df:

print(records_df.info())


And show missing values:

print(records_df.isnull().sum())

Then we do the same for inherited_df:

print(inherited_df.info())

And explore missing values for inherited houses. There are none here!

print(inherited_df.isnull().sum())

Great, now we have taken a first look at the date and can start doing some initial cleaning.

# Section 3: Cleaning - Data types mismatch

At this point, we spot that the data types vary between the records_df dataset and inherited_df dataset on numerous variables. Is this a problem?

Well, discrepancies between the data types in records_df and inherited_df for the same variables can cause issues down the line, especially when building and using predictive models.

Most machine learning algorithms require the input data to be in a consistent format. If a variable is represented as an integer in one dataset and as a float in another, the algorithm may get confused and produce unreliable results.

To address this, we should make sure that the same variables have the same data types in both datasets. Given that these variables represent counts (i.e., the number of square feet, the number of bedrooms), it would make more sense for them to be integers.

First we take a look at a sample of the data (top 20 rows), and see that the floats seem to be whole numbers. 

Therefore, in the cases where a variable is a float in one dataset and an integer in the other, it will be converted into an integer in both.

print(records_df.head(20))

print(inherited_df.head())

So we convert the variables listed below from the records_df dataframe to integers. Please note that the variable "GarageYrBlt" was also converted from a float to an integer in both datasets. This was not because of a data type mismatch, but because it made more sense to have the variable as an integer, as year values typically represent whole numbers and don't usually involve decimal points.

Given that there are missing values in some of the columns that we want to convert, we'll use the nullable integer type "Int64". Note that this type is case sensitive.

float_cols_records = ['2ndFlrSF', 'BedroomAbvGr', 'EnclosedPorch', 'LotFrontage', 'MasVnrArea', 'WoodDeckSF', 'GarageYrBlt']

for col in float_cols_records:
    records_df[col] = records_df[col].astype('Int64')


Now let's do the same for the inherited_df dataframe. In this case there are no missing values, so we can use the "int64" type.

float_cols_inherited = ['BsmtFinSF1', 'BsmtUnfSF', 'GarageArea', 'TotalBsmtSF', 'LotFrontage', 'MasVnrArea', 'GarageYrBlt']

for col in float_cols_inherited:
    inherited_df[col] = inherited_df[col].astype('int64')


Now we check the data types again to ensure the conversion was successful. First for the records_df:

print(records_df.info())


And here for inherited_df. Comparing them, we see that the data types for all variables match between the datasets, which will make analyses more reliable.

print(inherited_df.info())

# Section 4: Cleaning - Handle missing data values

Moving on, we proceed with handling missing data values. To refresh, here are the numbers of missing data in records_df. There were no missing data values in the inherited_df.

print(records_df.isnull().sum())

Looking at the missing values in records_df, columns 'EnclosedPorch' and 'WoodDeckSF' have a lot of missing values (over 80% of the data). Filling these missing values may not give us reliable data, so we choose the approach of dropping these columns. 

For other columns, we fill missing values with a reasonable strategy - using the median value for numerical columns and the most frequent value for categorical columns.

# Drop columns with too many missing values
records_df = records_df.drop(['EnclosedPorch', 'WoodDeckSF'], axis=1)
inherited_df = inherited_df.drop(['EnclosedPorch', 'WoodDeckSF'], axis=1)

# Fill missing values in numerical columns with the median
for col in ['2ndFlrSF', 'BedroomAbvGr', 'GarageYrBlt', 'LotFrontage', 'MasVnrArea']:
    records_df[col] = records_df[col].fillna(records_df[col].median())
    inherited_df[col] = inherited_df[col].fillna(inherited_df[col].median())

# Fill missing values in categorical columns with the most frequent value
for col in ['BsmtFinType1', 'GarageFinish']:
    records_df[col] = records_df[col].fillna(records_df[col].mode()[0])
    inherited_df[col] = inherited_df[col].fillna(inherited_df[col].mode()[0])

Then we print out the count of missing values in each column to confirm that there are no missing values left

print(records_df.isnull().sum())

And for inherited_df also here below, just for consistency. We can see here that the dropped variables are no longer visible.

print(inherited_df.isnull().sum())

Great! Now the dataset has no missing values and the data types are the same in both datasets.

---

## Save data

 Create a collection folder for records_df and manually drag and drop a copy of the raw dataset there for future use.

import os
try:
  os.makedirs(name='outputs/datasets/collection') # create outputs/datasets/collection folder
except Exception as e:
  print(e)

Now we create folders for the cleaned data and save it to new CSV files.

import os

# Define file paths
records_file_path = 'outputs/datasets/cleaned/house_prices_records_cleaned.csv'
inherited_file_path = 'outputs/datasets/cleaned/inherited_houses_cleaned.csv'

# Create the directories in the file path
os.makedirs(os.path.dirname(records_file_path), exist_ok=True)
os.makedirs(os.path.dirname(inherited_file_path), exist_ok=True)

# Save the CSV files
records_df.to_csv(records_file_path, index=False)
inherited_df.to_csv(inherited_file_path, index=False)

---

# Push files to Repo

Great! Now you can  push the changes to your GitHub Repo, using the Git commands (git add, git commit, git push)
